Refactor BrowseController to actions.
