+++ config/vufind/facets.ini
@@ -46,7 +46,7 @@ checkboxSections[] = CheckboxFacets
 
 ; This section is used to specify labels for facets that may be applied by parts
 ; of VuFind other than the facet lists defined in this file (for example, the
-; hierarchical browse of the BrowseController, or the Geographic Search).
+; hierarchical browse actions, or the Geographic Search).
 [ExtraFacetLabels]
 long_lat = "Geographic Search"
