+++ module/VuFind/config/module.config.php
@@ -168,11 +168,9 @@ $config = [
     'controllers' => [
         'factories' => [
             'VuFind\Controller\AlmaController' => 'VuFind\Controller\AbstractBaseFactory',
-            'VuFind\Controller\AlphabrowseController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\DeveloperSettingsController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\BlenderController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\Blender2Controller' => 'VuFind\Controller\AbstractBaseFactory',
-            'VuFind\Controller\BrowseController' => 'VuFind\Controller\AbstractBaseWithConfigFactory',
             'VuFind\Controller\BrowZineController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\CartController' => 'VuFind\Controller\CartControllerFactory',
             'VuFind\Controller\ChannelsController' => 'VuFind\Controller\ChannelsControllerFactory',
@@ -237,16 +235,12 @@ $config = [
         'aliases' => [
             'Alma' => 'VuFind\Controller\AlmaController',
             'alma' => 'VuFind\Controller\AlmaController',
-            'Alphabrowse' => 'VuFind\Controller\AlphabrowseController',
-            'alphabrowse' => 'VuFind\Controller\AlphabrowseController',
             'DeveloperSettings' => 'VuFind\Controller\DeveloperSettingsController',
             'developersettings' => 'VuFind\Controller\DeveloperSettingsController',
             'Blender' => 'VuFind\Controller\BlenderController',
             'blender' => 'VuFind\Controller\BlenderController',
             'Blender2' => 'VuFind\Controller\Blender2Controller',
             'blender2' => 'VuFind\Controller\Blender2Controller',
-            'Browse' => 'VuFind\Controller\BrowseController',
-            'browse' => 'VuFind\Controller\BrowseController',
             'BrowZine' => 'VuFind\Controller\BrowZineController',
             'browzine' => 'VuFind\Controller\BrowZineController',
             'Cart' => 'VuFind\Controller\CartController',
