+++ module/VuFind/src/VuFind/Action/Browse/AuthorAction.php
@@ -1,0 +1,69 @@
+<?php
+
+/**
+ * Browse authors action.
+ *
+ * PHP version 8
+ *
+ * Copyright (C) The National Library of Finland 2026.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, see
+ * <https://www.gnu.org/licenses/>.
+ *
+ * @category VuFind
+ * @package  Action
+ * @author   Ere Maijala <ere.maijala@helsinki.fi>
+ * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
+ * @link     https://vufind.org Main Site
+ */
+
+namespace VuFind\Action\Browse;
+
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+
+/**
+ * Browse authors action.
+ *
+ * @category VuFind
+ * @package  Action
+ * @author   Ere Maijala <ere.maijala@helsinki.fi>
+ * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
+ * @link     https://vufind.org Main Site
+ */
+class AuthorAction extends AbstractBrowseAction
+{
+    /**
+     * Browse authors.
+     *
+     * @param ServerRequestInterface $request  Server request
+     * @param ResponseInterface      $response Response
+     *
+     * @return ResponseInterface
+     */
+    public function action(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+    ): ResponseInterface {
+        $categoryList = [
+            'alphabetical' => 'By Alphabetical',
+            'lcc'          => 'By Call Number',
+            'topic'        => 'By Topic',
+            'genre'        => 'By Genre',
+            'region'       => 'By Region',
+            'era'          => 'By Era',
+        ];
+
+        return $this->performBrowse('Author', $categoryList, true);
+    }
+}
